requirements.txt
asgiref==3.2.3
Django==3.0.7
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.0.0
django-cors-headers==3.3.0
inflection==0.3.1
numpy==1.17.4
pandas==0.25.3
postgres==3.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.8
virtualenvwrapper-win==1.2.5
xlwt==1.3.0

